requirements.txt
langchain
langchain_community
langchain_core
langchain_huggingface
langchain-text-splitter
langchain_groq
python-dotenv
streamlit
pymupdf
pypdf
faiss-cpu
python-docx
